revert: remove index idx_server_namespace from attributes, versioned_attributes
This adds back the idx_server_namespace index.

The change being reverted was made on the assumption that both the ServerID
and the ComponentID are to be declared for each component attribute.

Based on the discussion in the PR https://github.com/metal-toolbox/hollow-serverservice/pull/101#discussion_r915225582
either the ServerID or the Server ComponentID should be set.

reverts commit:

commit 988350ab683150a42c1393f39aa4f2a75f7d4be2
Author: Joel Rebello <jrebello@packet.com>
Date:   Wed Jul 6 21:56:51 2022 +0200

    db/migrations: remove index idx_server_namespace from attributes, versioned_attributes

    This index prevents multiple component attributes that reference the same server ID
    from being inserted. Purging this index until a better work around is determined.

    hollow client received a server error - response code: 500, message: datastore
    error, details: failed to insert into foreign table: models: unable to insert
    into attributes: pq: duplicate key value violates unique constraint "idx_server_namespace"